+++ src/neo/SmartContract/Native/ContractMethodAttribute.cs
@@ -6,13 +6,13 @@ namespace Neo.SmartContract.Native
     internal class ContractMethodAttribute : Attribute
     {
         public string Name { get; set; }
-        public long Price { get; }
+        public uint Price { get; }
         public ContractParameterType ReturnType { get; }
         public ContractParameterType[] ParameterTypes { get; set; } = Array.Empty<ContractParameterType>();
         public string[] ParameterNames { get; set; } = Array.Empty<string>();
         public bool SafeMethod { get; set; } = false;
 
-        public ContractMethodAttribute(long price, ContractParameterType returnType)
+        public ContractMethodAttribute(uint price, ContractParameterType returnType)
         {
             this.Price = price;
             this.ReturnType = returnType;
